Initial Commit
Basically this is a copy of `opennms-pagerduty-plugin`.

There are a lot of pieces that don't apply to the StatusPage plugin:
- code under package `web` and `webhook`.
- code under package `client`.

Removal of these will happen after solving the more fundamental problem of obtaining a clean build.

Changes have been made to (attempt) support JDK 11 and the version of Karaf used by Horizon (4.3.2 at time of writing).